blooddonation/requirements.txt
africastalking==1.2.4
appdirs==1.4.3
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
blinker==1.4
brow==0.0.3
captain==3.0.1
certifi==2022.9.24
chardet==3.0.4
Click==7.0
colorama==0.4.3
contextlib2==21.6.0
cryptography==2.8

dbus-python==1.2.16

distlib==0.3.0


dj-database-url==1.0.0
Django==4.1.2
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0

entrypoints==0.3
exceptiongroup==1.0.0rc9
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
h11==0.14.0
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
keras==2.10.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
npm==0.1.1
numpy==1.23.4
oauthlib==3.1.0
olefile==0.46
optional-django==0.1.0
outcome==1.2.0
packaging==20.3
pandas==1.5.1
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.8.6
pycairo==1.16.2
pycups==1.9.73
pydantic==1.10.2
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2022.5
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
schema==0.7.5
SecretStorage==2.3.1
selenium==4.5.0
self-py==0.0.1a0
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.2.4
systemd-python==234
toolz==0.12.0
trio==0.22.0
trio-websocket==0.9.2
ttystatus==0.38
typing_extensions==4.4.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
whitenoise==6.2.0
wishlist==0.5.3
wsproto==1.2.0
xkit==0.0.0
zipp==1.0.0
